+++ public/components/common/flyouts/wz-flyout.tsx
@@ -11,33 +11,12 @@
  */
 
 import React from 'react';
-import { EuiFlyout, EuiOutsideClickDetector } from '@elastic/eui';
+import { EuiFlyout } from '@elastic/eui';
 
 export const WzFlyout = ({ children, flyoutProps = {}, onClose }) => {
-  const closeFlyout = (ev) => {
-    // Clicking on the flyout or on the flyout selector should not close the flyout.
-    if (
-      ev &&
-      ev.path.some(
-        (element) =>
-          element.classList?.contains('euiFlyout') || element.classList?.contains('euiPanel')
-      )
-    ) {
-      return;
-    }
-    onClose();
-  };
-
   return (
-    <EuiOutsideClickDetector onOutsideClick={closeFlyout}>
-      {/*
-        As the version of Elastic EUI (v34.6.0) has a bug in the EuiOverlayMask component,
-        maskProps is added to avoid the closing of the overlay by the native function, as it contains the bug
-        This bug is fixed in the Elastic EUI version 36.0.0
-      */}
-      <EuiFlyout onClose={onClose} maskProps={{ onClick: () => {} }} {...flyoutProps}>
-        {children}
-      </EuiFlyout>
-    </EuiOutsideClickDetector>
-  );
+    <EuiFlyout onClose={onClose} maskProps={{ onClick: () => {} }} {...flyoutProps} outsideClickCloses={true}>
+      {children}
+    </EuiFlyout>
+  )
 };